+++ lib/utils/message.dart
@@ -4,6 +4,8 @@ class Message {
   final String nickname;
   final DateTime time;
   final bool isMe;
+  bool isReaded;
 
-  Message(this.message, this.from, this.nickname, this.time, this.isMe);
+  Message(this.message, this.from, this.nickname, this.time, this.isMe,
+      this.isReaded);
 }