+++ lib/utils/session.dart
@@ -8,6 +8,7 @@ import 'package:openvidu_flutter/utils/custom_websocket.dart';
 import 'package:openvidu_flutter/utils/message.dart';
 
 typedef OnNotifySetRemoteMediaStreamEvent = void Function(String id);
+typedef OnAddRemoteParticipantEvent = void Function(String id);
 typedef OnRemoveRemoteParticipantEvent = void Function(String id);
 typedef OnMessageReceivedEvent = void Function(Message msg);
 
@@ -18,11 +19,14 @@ class Session {
   String token;
   List<Message> messages = [];
   OnNotifySetRemoteMediaStreamEvent? onNotifySetRemoteMediaStream;
+  OnAddRemoteParticipantEvent? onAddRemoteParticipant;
   OnRemoveRemoteParticipantEvent? onRemoveRemoteParticipant;
   final StreamController<Message> _messageStreamController =
       StreamController<Message>.broadcast();
 
   Stream<Message> get messageStream => _messageStreamController.stream;
+
+  int get unreadMessages => messages.where((msg) => !msg.isReaded).length;
 
   final List<Map<String, List<String>>> iceServersDefault = [
     {
@@ -180,8 +184,15 @@ class Session {
 
   void addMessageReceived(Map<String, dynamic> params) {
     var data = jsonDecode(params['data']);
-    var msg = Message(data['message'], params['from'], data['nickname'],
-        DateTime.now(), localParticipant?.participantName == data['nickname']);
+    final isMe = localParticipant?.participantName == data['nickname'];
+    var msg = Message(
+      data['message'],
+      params['from'],
+      data['nickname'],
+      DateTime.now(),
+      isMe,
+      isMe,
+    );
     messages.add(msg);
     _messageStreamController.add(msg);
   }